+++ src/dashboard/pages/Category/Category.tsx
@@ -1,16 +1,15 @@
 import React, { useEffect } from 'react';
 
 import { Layout, Pagination } from 'antd';
-import HeaderTitle from '../../components/HeaderTitle';
-import TableCategory from './TableCategory';
-import { useTitle } from '../../../hooks/useTitle';
-import { categoryTypeActions } from '../../../redux/slice/categoryTypeSlice';
 import { useDispatch, useSelector } from 'react-redux';
+import { useTitle } from '../../../hooks/useTitle';
 import {
   categoryActions,
   categorySelector,
   categoryState,
 } from '../../../redux/slice/categorySlice';
+import HeaderTitle from '../../components/HeaderTitle';
+import TableCategory from './TableCategory';
 
 const { Content } = Layout;
 
@@ -18,10 +17,6 @@ const Category: React.FC = () => {
   const dispatch = useDispatch();
   const { page, categories, pageSize }: categoryState =
     useSelector(categorySelector);
-
-  useEffect(() => {
-    dispatch(categoryTypeActions.getAllCategoryType({}));
-  }, [dispatch]);
 
   useEffect(() => {
     dispatch(
