+++ src/dashboard/pages/Category/FormCategory/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FormCategory';
