--- src/dashboard/pages/Category/ModalCategory/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ModalCategory';
